refactor(arbitration): split player state from player strategy
