+++ onchain/permissionless-arbitration/offchain/player/dishonest_player.lua
@@ -3,11 +3,12 @@ package.path = package.path .. ";/opt/cartesi/lib/lua/5.4/?.lua"
 package.path = package.path .. ";./offchain/?.lua"
 package.cpath = package.cpath .. ";/opt/cartesi/lib/lua/5.4/?.so"
 
-local Player = require "player.honest_strategy"
+local Player = require "player.state"
 local Client = require "blockchain.client"
 local Hash = require "cryptography.hash"
 
 local time = require "utils.time"
+local strategy = require "player.strategy"
 
 local player_index = tonumber(arg[1])
 local tournament = arg[2]
@@ -21,6 +22,7 @@ do
 end
 
 while true do
-    if p:react() then break end
+    p:fetch()
+    if strategy.react_honestly(p) then break end
     time.sleep(1)
 end
